Change the way of authenticating with BCC when making requests to it
We now use an x-mss-signature plus a couple of other headers to
authenticate that a given request is from MSS. The signature is
created with MSS's private key and verified with MSS public key.
All other routes in BCC that should only be accessed by MSS have
as similar authentication mechanism.